Replay write-ahead log on load
Replay the write-ahead log when the database is loaded.
This implements durability as demanded by issue #6.

What is left TODO:

Currently all view- and storage drivers truncate their data files
on load and after that all transactions are replayed from the log.
This should be improved in future commits.

We also do not re-submit transactions that were submitted but
not commited. This should be fixed.