Introduce SequenceIds
A SequenceId is a unique id number assigned to each
commited transaction. SequenceIds are used to keep track of when
changes to the database happen; they are always increasing
so they can be used to check which version of a document is newer.
You can think of them as "revision ids".

This is part of a solution to issue #5.